+++ bugzoo/client/container.py
@@ -181,6 +181,23 @@ class ContainerManager(object):
             return TestOutcome.from_dict(r.json())
         self.__api.handle_erroneous_response(r)
 
+    def coverage(self,
+                 container: Container
+                 ) -> TestSuiteCoverage:
+        uid = container.uid
+        logger.info("Fetching coverage information for container: %s",
+                    uid)
+        r = self.__api.get('containers/{}/coverage'.format(uid))
+        if r.status_code == 200:
+            jsn = r.json()
+            coverage = TestSuiteCoverage.from_dict(jsn)  # type: ignore
+            logger.info("Fetched coverage information for container: %s",
+                        uid)
+            return coverage
+        logger.error("Failed to fetch coverage information for container: %s",
+                     uid)
+        self.__api.handle_erroneous_response(r)
+
     def exec(self,
              container: Container,
              command: str,