+++ bugzoo/mgr/coverage.py
@@ -130,20 +130,25 @@ class CoverageManager(object):
 
     def coverage(self,
                  container: Container,
-                 tests: List[TestCase],
+                 tests: Optional[List[TestCase]] = None,
                  files_to_instrument: List[str] = None,
                  ) -> TestSuiteCoverage:
         """
         Uses a provided container to compute line coverage information for a
         given list of tests.
         """
-        assert tests != []
+        if tests is None:
+            bug = self.__installation.bugs[container.bug]
+            _tests = bug.tests
+        else:
+            assert tests is not []
+            _tests = tests
 
         self.instrument(container,
                         files_to_instrument=files_to_instrument)
 
         cov = {}
-        for test in tests:
+        for test in _tests:
             print("generating coverage: {}".format(test))
             outcome = self.__installation.containers.execute(container, test)
             filelines = self.extract(container,